requirements.txt

Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Faker==19.3.0
